requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
confusable-homoglyphs==3.3.1
crispy-bootstrap5==2024.10
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.0.2
django-cors-headers==4.5.0
django-crispy-forms==2.3
django-registration==3.4
django-rest-auth==0.9.5
django-webpack-loader==3.1.1
djangorestframework==3.15.2
idna==3.10
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.3.0
gunicorn==20.0.0